Rename footer -> Footer
