+++ src/App.jsx
@@ -1,7 +1,7 @@
 import "./App.css";
 import { Outlet } from "react-router-dom";
 import Header from "./components/header/Header";
-import Footer from "./components/footer/Footer";
+import Footer from "./components/Footer/Footer";
 import { useLocation } from "react-router-dom";
 import { AuthProvider } from "./utils/authContext";
 import { useEffect, useRef, useState } from "react";
